Extract Manifest for building a map of the ZIP
We need a map of byteranges within the ZIP to
allow for partial ZIP builds. Therefore it
makes sense to expand the size estimator to
include a generator for those maps. This is
easier to do if we just build the map in a
separate class, and then use that class
from the size estimator, preserving the API