Inline assertions for cdir checks
We are reading the contents of a buffer _per byte_, in
a linear fashion. It is important, when doing so, that
the order of reads from the buffer corresponds 1 to 1
to how the test assertions are laid out, because humans
are notoriously bad at following control flows when very
literal value reads are involved.

So when we perform these assertions, it is a good idea
to literally specify _every value_ that we check for, even
though it involves copy-paste and repeated code. But when
time comes to evaluate changes and figure out what actually
goes wrong somewhere, we can scan the code in a linear fashion
as opposed to jumping from the example to the check method, from
the assertion in the check method to a conditional variable, then
back to the assertion in the example etc.

This is _exactly_ the situation where DRY does _not_ apply.
Trust me.