Fix the bug with reading Zip64 extra fields
According to the APPNOTE, the Zip64 extra fields can be set
only if the corresponding "standard" field in the central dir
entry is blanked to the overflow value (either 0xFFFF or
0xFFFFFFFFF). Therefore we are not guaranteed to have all
these fields present in the extra. Before reading, we need
to check _which_ fields in the central directory entry that
we have already ingested have overflow values, and _only_ read
those specific fields.