Use a "tuple array" to pack values
In theory, if we do it the way we used to we
need to allocate one Array for each value written.
If we allocate an Array for the entire block of
values, and then just split it into 2, we might
be able to reduce object churn. We can also then
use only 1 pack() call to process the entire block
of data as opposed to packing individually.