Fix quadratic perf problem when uniqing filenames (#14)
We check for unique filenames using a `.any?` on an
Array. It is going to take longer as the array grows,
since `any?` needs to do an array scan. This leads to
a quadratic performance regression when file counts in
the ZIP become large.

Without this patch, writing 40000 fake files into
a null writer takes _minutes_.