Speed up StreamCRC32 using a buffer
When doing a lot of small writes, combining the CRC32
checksums becomes expeinsive (up to 20% of wall clock
time was spent on a test with a million writes). We
introduce a buffer to ship bytes to Zlib in larger
chunks, therefore benefitting from speedups within
zlib itself and doing less bridging calls to boot.