Add Streamer#output_enum to support deferred writes (#49)
The pattern that is required for output of ZIPs from Rack and the pattern for deferred output (see #45 ) are actually the same. All we need to do to facilitate both is to provide an eachable enumerator for the output, so that the calling code can have control.

This PR strips the RackBody to the minimum and replaces it with a more generic output enumerator object that supports this usage pattern. I also removed some #close methods which were no-ops anyway, so a major version bump will probably be in order to support this.