Optimize CRC32 allocations for IOs (#57)
We can reuse a string buffer across multiple reads, thereby
reducing allocations and GC churn. Also we can raise the
buffer size as reading the file is usually more expensive than
computing the CRC. And to top it off on recent Rubies we can
pre-cap our String to the right size and it will not have to
be resized on every alteration, sparing cycles. It uses two
semi-obscure features:

**Capacity preallocation**

The constructor for `String` accepts an optional `capacity`
keyword argument, since
https://bugs.ruby-lang.org/issues/12024 The argument can be
used to pre-size the String to the specific number of bytes.
If the string gets mutated, and the new string contents is not
larger than the `capacity` of the string, the string will not
need resizing - meaning we never need to realloc for the
string contents. We must run on older Rubies too so that
optimization is conditional.

**IO#read with the buffer argument**

All `IO#read` calls accept an optional `outbuf` argument which
should be a String. IO then replaces the contents of that
string instead of allocating a new one, and returns it. This
makes it possible to reuse one and the same string object
throughout multiple `read` calls. See
https://apidock.com/ruby/IO/read for more. Since we only read
as many bytes as we set in `capacity`, we won't be resizing
the string either.

We use a larger buffer size as this method has different
characteristics than the one we use for stream-writes from
the Streamer.

**Benchmarks**

https://gist.github.com/julik/ada7229c8d85290aad3c72572c368f01