Add PathSet and check for file/directory conflicts
We have been having a number of situations where
an implicitly created directory in longer paths
would "occupy" a path component that would then
be taken up by a file. With some applications it
is simply not possible to extract such an archive,
and also the filesystem structure it would create
is not really defined.

To avoid this we should verify the paths added
during the archive creation - and we need to do
it when every entry gets added to the internal
entry list, at the earliest possible opportunity.

It also shows that our uniquify calls to avoid
duplicate filenames lead to ambiguous situations,
so we make this optional with a default of "true".

We keep the option enabled by default to make sure
we do not break zipline which relies on us doing
the uniqueness fixing for them (we asked for it).

So even though we can fix it in our own output,
it is not reasonable to expect all of our users to
do the same, _and_ there are guarantees of semver.