Adopt MIT - Hippocratic license (#61)
As @bermannoah noticed in the format_parser
pull request it is a good idea for us to
adopt a more restrictive license.

This makes the license zip_tricks is under
more restrictive than it used to be. To
limit impact on downstream dependencies
such as xlsxtream I opted to increment the
major version of the gem so that its
update will be more visible and the dependents
can elect to stay on 4.x version tree.

The 4.x will stay under the basic MIT license
so that we do not suddenly de-license current
users of the software.