Fix extended timestamp extra field output (#67)
Upon closer inspection of why Rubyzip 2.1 cannot work with our output it turns out that we write the timestamp extra fields incorrectly. To indicate that only mtime is present, the last bit of the flag byte needs to be set. We were erroneously setting the highest bit instead.

Since Rubyzip 2 attempts to read the extra timestamp field, and does it - it doesn't find the flag bit which instructs it to read the mtime from the extra field. Rubyzip does not parse the "standard" time field if the timestamp extra field is present at all - so both the standard time for the entry and the time from the extra field end up being nil.

Closes #65