Tweak CRC32 buffer size, preallocate on all rubies
Benchmarks show that no speedup is gained above a CRC32 buffer size of
64K, so using a larger buffer is wasting memory and increasing pressure
on the garbage collector.

Instead of using String.new with capacity keyword, we simply repeat a
null byte string. This is further optimized on MRI 2.5.0 and later by
being implemented using calloc, so no zeroing has to be done.