Remove complexity_assert development dependency
This was used by the streamer spec to assert linear complexity, but
causes a huge amount of array allocations with 300 MB memory usage
and takes up about 40% of the entire test suite runtime for 1 spec.

Instead replace it with an inline implementation that asserts linear
complexity with 75% confidence interval.

Depending on how much performance fluctuation is seen in CI, the
confidence interval might have to be lowered.